style: fix clippy warnings so `-D warnings` passes cleanly
Move `save_config` above the test module (items_after_test_module) and
drop a needless i32 cast in the OSD test (unnecessary_cast). These were
the last two lint warnings; clippy is now fully clean under -D warnings.

+++ src/osd.rs
@@ -455,7 +455,7 @@ mod tests {
         // Canvas is large enough that the (now bigger) legend box always fits
         // entirely in the chosen corner quadrant.
         let size = 600;
-        let mid = (size / 2) as i32;
+        let mid = size / 2;
         let corners = [
             ("top-left", (400, 400)),
             ("top-right", (100, 400)),
